# jnp.issubdtype: refactor for better performance