#17740 led to a ~10% regression in our benchmark for simple `device_put`. Via line profiling, I found that the issue came from the fact that `dtypes.issubdtype(x, dtypes.extended)` is called frequently within JAX's internals (e.g. five times for `jax.device_put(np.arange(4))`), and #17740 had made this marginally more expensive by not handling this case up front.

This change refactors the function to do this handling up front, and also make the normal path somewhat more efficient.

### Example benchmarks

Before #17740:
```python
In [1]: import numpy as np
   ...: import jax
   ...: x = np.array(1, dtype=np.int32)
   ...: %timeit jax.device_put(x).block_until_ready()
22.7 µs ± 115 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

After #17740:
```python
In [1]: import numpy as np
   ...: import jax
   ...: x = np.array(1, dtype=np.int32)
   ...: %timeit jax.device_put(x).block_until_ready()
27.6 µs ± 286 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

With this PR:
```python
In [1]: import numpy as np
   ...: import jax
   ...: x = np.array(1, dtype=np.int32)
   ...: %timeit jax.device_put(x).block_until_ready()
23.2 µs ± 99.2 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```